+++ src/Config/MspGpioMap.php
@@ -39,6 +39,7 @@ final class MspGpioMap
         107 => 'FreqWakeUp',
         108 => 'notifMode',
         109 => 'notifCategories',
+        110 => 'resetMode',
         111 => 'ServoModeAuto',
     ];