+++ tests/Config/MspGpioMapCoherenceTest.php
@@ -37,6 +37,7 @@ class MspGpioMapCoherenceTest extends TestCase
         107 => 'FreqWakeUp',
         108 => 'notifMode',
         109 => 'notifCategories',
+        110 => 'resetMode',
         111 => 'ServoModeAuto',
     ];